Extract validator expression helpers
