+++ crates/rulemorph/src/validator/expr.rs
@@ -1,0 +1,606 @@
+use std::collections::HashSet;
+
+use crate::error::ErrorCode;
+use crate::model::{Expr, ExprChain, ExprOp};
+use crate::path::PathToken;
+
+use super::ValidationCtx;
+use super::expr_args::{
+    validate_lookup_args, validate_lookup_args_chain, validate_path_arg, validate_path_array_arg,
+};
+use super::op_inventory::{element_expr_scope, is_valid_op};
+use super::refs::validate_ref;
+use super::scope::LocalScope;
+
+pub(super) fn validate_expr(
+    expr: &Expr,
+    base_path: &str,
+    produced_targets: &HashSet<Vec<PathToken>>,
+    ctx: &mut ValidationCtx<'_>,
+    scope: LocalScope,
+) {
+    match expr {
+        Expr::Ref(expr_ref) => validate_ref(expr_ref, base_path, produced_targets, ctx, scope),
+        Expr::Op(expr_op) => validate_op(expr_op, base_path, produced_targets, ctx, scope),
+        Expr::Chain(expr_chain) => {
+            validate_chain(expr_chain, base_path, produced_targets, ctx, scope)
+        }
+        Expr::Literal(_) => {}
+    }
+}
+
+fn validate_chain(
+    expr_chain: &ExprChain,
+    base_path: &str,
+    produced_targets: &HashSet<Vec<PathToken>>,
+    ctx: &mut ValidationCtx<'_>,
+    scope: LocalScope,
+) {
+    if expr_chain.chain.is_empty() {
+        ctx.push(
+            ErrorCode::InvalidExprShape,
+            "expr.chain must be a non-empty array",
+            format!("{}.chain", base_path),
+        );
+        return;
+    }
+
+    for (index, item) in expr_chain.chain.iter().enumerate() {
+        let item_path = format!("{}.chain[{}]", base_path, index);
+        if index == 0 {
+            validate_expr(item, &item_path, produced_targets, ctx, scope);
+            continue;
+        }
+
+        match item {
+            Expr::Op(expr_op) => {
+                validate_chain_op(expr_op, &item_path, produced_targets, ctx, scope);
+            }
+            _ => {
+                ctx.push(
+                    ErrorCode::InvalidExprShape,
+                    "expr.chain items after first must be op",
+                    item_path,
+                );
+            }
+        }
+    }
+}
+
+fn validate_chain_op(
+    expr_op: &ExprOp,
+    base_path: &str,
+    produced_targets: &HashSet<Vec<PathToken>>,
+    ctx: &mut ValidationCtx<'_>,
+    scope: LocalScope,
+) {
+    if !is_valid_op(&expr_op.op) {
+        ctx.push(
+            ErrorCode::UnknownOp,
+            "expr.op is not supported",
+            format!("{}.op", base_path),
+        );
+    }
+
+    let args_len = expr_op.args.len() + 1;
+    match expr_op.op.as_str() {
+        "trim" | "lowercase" | "uppercase" | "to_string" | "len" | "not" => {
+            if args_len != 1 {
+                ctx.push(
+                    ErrorCode::InvalidArgs,
+                    "expr.args must contain exactly one item",
+                    format!("{}.args", base_path),
+                );
+            }
+        }
+        "replace" => {
+            if !(3..=4).contains(&args_len) {
+                ctx.push(
+                    ErrorCode::InvalidArgs,
+                    "expr.args must contain three or four items",
+                    format!("{}.args", base_path),
+                );
+            }
+        }
+        "split" => {
+            if args_len != 2 {
+                ctx.push(
+                    ErrorCode::InvalidArgs,
+                    "expr.args must contain exactly two items",
+                    format!("{}.args", base_path),
+                );
+            }
+        }
+        "pad_start" | "pad_end" => {
+            if !(2..=3).contains(&args_len) {
+                ctx.push(
+                    ErrorCode::InvalidArgs,
+                    "expr.args must contain two or three items",
+                    format!("{}.args", base_path),
+                );
+            }
+        }
+        "lookup" | "lookup_first" => {
+            validate_lookup_args_chain(expr_op, base_path, ctx);
+        }
+        "merge" | "deep_merge" => {
+            if args_len < 2 {
+                ctx.push(
+                    ErrorCode::InvalidArgs,
+                    "expr.args must contain at least two items",
+                    format!("{}.args", base_path),
+                );
+            }
+        }
+        "get" => {
+            if args_len != 2 {
+                ctx.push(
+                    ErrorCode::InvalidArgs,
+                    "expr.args must contain exactly two items",
+                    format!("{}.args", base_path),
+                );
+            } else {
+                validate_path_arg(&expr_op.args[0], &format!("{}.args[0]", base_path), ctx);
+            }
+        }
+        "pick" | "omit" => {
+            if args_len != 2 {
+                ctx.push(
+                    ErrorCode::InvalidArgs,
+                    "expr.args must contain exactly two items",
+                    format!("{}.args", base_path),
+                );
+            } else {
+                let allow_terminal_index = expr_op.op == "pick";
+                validate_path_array_arg(
+                    &expr_op.args[0],
+                    &format!("{}.args[0]", base_path),
+                    allow_terminal_index,
+                    ctx,
+                );
+            }
+        }
+        "keys" | "values" | "entries" | "object_flatten" | "object_unflatten" => {
+            if args_len != 1 {
+                ctx.push(
+                    ErrorCode::InvalidArgs,
+                    "expr.args must contain exactly one item",
+                    format!("{}.args", base_path),
+                );
+            }
+        }
+        "from_entries" => {
+            if !(1..=2).contains(&args_len) {
+                ctx.push(
+                    ErrorCode::InvalidArgs,
+                    "expr.args must contain one or two items",
+                    format!("{}.args", base_path),
+                );
+            }
+        }
+        "map" | "filter" | "flat_map" | "group_by" | "key_by" | "partition" | "distinct_by"
+        | "find" | "find_index" => {
+            if args_len != 2 {
+                ctx.push(
+                    ErrorCode::InvalidArgs,
+                    "expr.args must contain exactly two items",
+                    format!("{}.args", base_path),
+                );
+            }
+        }
+        "flatten" => {
+            if !(1..=2).contains(&args_len) {
+                ctx.push(
+                    ErrorCode::InvalidArgs,
+                    "expr.args must contain one or two items",
+                    format!("{}.args", base_path),
+                );
+            }
+        }
+        "take" | "drop" | "chunk" | "index_of" | "contains" | "reduce" => {
+            if args_len != 2 {
+                ctx.push(
+                    ErrorCode::InvalidArgs,
+                    "expr.args must contain exactly two items",
+                    format!("{}.args", base_path),
+                );
+            }
+        }
+        "slice" => {
+            if !(2..=3).contains(&args_len) {
+                ctx.push(
+                    ErrorCode::InvalidArgs,
+                    "expr.args must contain two or three items",
+                    format!("{}.args", base_path),
+                );
+            }
+        }
+        "zip" => {
+            if args_len < 2 {
+                ctx.push(
+                    ErrorCode::InvalidArgs,
+                    "expr.args must contain at least two items",
+                    format!("{}.args", base_path),
+                );
+            }
+        }
+        "zip_with" => {
+            if args_len < 3 {
+                ctx.push(
+                    ErrorCode::InvalidArgs,
+                    "expr.args must contain at least three items",
+                    format!("{}.args", base_path),
+                );
+            }
+        }
+        "unzip" | "unique" | "sum" | "avg" | "min" | "max" => {
+            if args_len != 1 {
+                ctx.push(
+                    ErrorCode::InvalidArgs,
+                    "expr.args must contain exactly one item",
+                    format!("{}.args", base_path),
+                );
+            }
+        }
+        "sort_by" => {
+            if !(2..=3).contains(&args_len) {
+                ctx.push(
+                    ErrorCode::InvalidArgs,
+                    "expr.args must contain two or three items",
+                    format!("{}.args", base_path),
+                );
+            }
+        }
+        "fold" => {
+            if args_len != 3 {
+                ctx.push(
+                    ErrorCode::InvalidArgs,
+                    "expr.args must contain exactly three items",
+                    format!("{}.args", base_path),
+                );
+            }
+        }
+        "+" | "*" | "and" | "or" => {
+            if args_len < 2 {
+                ctx.push(
+                    ErrorCode::InvalidArgs,
+                    "expr.args must contain at least two items",
+                    format!("{}.args", base_path),
+                );
+            }
+        }
+        "-" | "/" | "to_base" | "==" | "!=" | "<" | "<=" | ">" | ">=" | "~=" => {
+            if args_len != 2 {
+                ctx.push(
+                    ErrorCode::InvalidArgs,
+                    "expr.args must contain exactly two items",
+                    format!("{}.args", base_path),
+                );
+            }
+        }
+        "round" => {
+            if !(1..=2).contains(&args_len) {
+                ctx.push(
+                    ErrorCode::InvalidArgs,
+                    "expr.args must contain one or two items",
+                    format!("{}.args", base_path),
+                );
+            }
+        }
+        "date_format" => {
+            if !(2..=4).contains(&args_len) {
+                ctx.push(
+                    ErrorCode::InvalidArgs,
+                    "expr.args must contain two to four items",
+                    format!("{}.args", base_path),
+                );
+            }
+        }
+        "to_unixtime" => {
+            if !(1..=3).contains(&args_len) {
+                ctx.push(
+                    ErrorCode::InvalidArgs,
+                    "expr.args must contain one to three items",
+                    format!("{}.args", base_path),
+                );
+            }
+        }
+        _ => {}
+    }
+
+    let expr_scope = element_expr_scope(&expr_op.op, true, expr_op.args.len(), scope);
+    for (index, arg) in expr_op.args.iter().enumerate() {
+        let arg_path = format!("{}.args[{}]", base_path, index);
+        let arg_scope = match expr_scope {
+            Some((expr_index, expr_scope)) if expr_index == index => expr_scope,
+            _ => scope,
+        };
+        validate_expr(arg, &arg_path, produced_targets, ctx, arg_scope);
+    }
+}
+
+fn validate_op(
+    expr_op: &ExprOp,
+    base_path: &str,
+    produced_targets: &HashSet<Vec<PathToken>>,
+    ctx: &mut ValidationCtx<'_>,
+    scope: LocalScope,
+) {
+    if !is_valid_op(&expr_op.op) {
+        ctx.push(
+            ErrorCode::UnknownOp,
+            "expr.op is not supported",
+            format!("{}.op", base_path),
+        );
+    }
+
+    if expr_op.args.is_empty() {
+        ctx.push(
+            ErrorCode::InvalidArgs,
+            "expr.args must be a non-empty array",
+            format!("{}.args", base_path),
+        );
+    }
+
+    match expr_op.op.as_str() {
+        "trim" | "lowercase" | "uppercase" | "to_string" | "len" => {
+            if expr_op.args.len() != 1 {
+                ctx.push(
+                    ErrorCode::InvalidArgs,
+                    "expr.args must contain exactly one item",
+                    format!("{}.args", base_path),
+                );
+            }
+        }
+        "replace" => {
+            if !(3..=4).contains(&expr_op.args.len()) {
+                ctx.push(
+                    ErrorCode::InvalidArgs,
+                    "expr.args must contain three or four items",
+                    format!("{}.args", base_path),
+                );
+            }
+        }
+        "split" => {
+            if expr_op.args.len() != 2 {
+                ctx.push(
+                    ErrorCode::InvalidArgs,
+                    "expr.args must contain exactly two items",
+                    format!("{}.args", base_path),
+                );
+            }
+        }
+        "pad_start" | "pad_end" => {
+            if !(2..=3).contains(&expr_op.args.len()) {
+                ctx.push(
+                    ErrorCode::InvalidArgs,
+                    "expr.args must contain two or three items",
+                    format!("{}.args", base_path),
+                );
+            }
+        }
+        "lookup" | "lookup_first" => {
+            validate_lookup_args(expr_op, base_path, ctx);
+        }
+        "merge" | "deep_merge" => {
+            if expr_op.args.len() < 2 {
+                ctx.push(
+                    ErrorCode::InvalidArgs,
+                    "expr.args must contain at least two items",
+                    format!("{}.args", base_path),
+                );
+            }
+        }
+        "get" => {
+            if expr_op.args.len() != 2 {
+                ctx.push(
+                    ErrorCode::InvalidArgs,
+                    "expr.args must contain exactly two items",
+                    format!("{}.args", base_path),
+                );
+            } else {
+                validate_path_arg(&expr_op.args[1], &format!("{}.args[1]", base_path), ctx);
+            }
+        }
+        "pick" | "omit" => {
+            if expr_op.args.len() != 2 {
+                ctx.push(
+                    ErrorCode::InvalidArgs,
+                    "expr.args must contain exactly two items",
+                    format!("{}.args", base_path),
+                );
+            } else {
+                let allow_terminal_index = expr_op.op == "pick";
+                validate_path_array_arg(
+                    &expr_op.args[1],
+                    &format!("{}.args[1]", base_path),
+                    allow_terminal_index,
+                    ctx,
+                );
+            }
+        }
+        "keys" | "values" | "entries" | "object_flatten" | "object_unflatten" => {
+            if expr_op.args.len() != 1 {
+                ctx.push(
+                    ErrorCode::InvalidArgs,
+                    "expr.args must contain exactly one item",
+                    format!("{}.args", base_path),
+                );
+            }
+        }
+        "from_entries" => {
+            if !(1..=2).contains(&expr_op.args.len()) {
+                ctx.push(
+                    ErrorCode::InvalidArgs,
+                    "expr.args must contain one or two items",
+                    format!("{}.args", base_path),
+                );
+            }
+        }
+        "map" | "filter" | "flat_map" | "group_by" | "key_by" | "partition" | "distinct_by"
+        | "find" | "find_index" => {
+            if expr_op.args.len() != 2 {
+                ctx.push(
+                    ErrorCode::InvalidArgs,
+                    "expr.args must contain exactly two items",
+                    format!("{}.args", base_path),
+                );
+            }
+        }
+        "flatten" => {
+            if !(1..=2).contains(&expr_op.args.len()) {
+                ctx.push(
+                    ErrorCode::InvalidArgs,
+                    "expr.args must contain one or two items",
+                    format!("{}.args", base_path),
+                );
+            }
+        }
+        "take" | "drop" | "chunk" | "index_of" | "contains" | "reduce" => {
+            if expr_op.args.len() != 2 {
+                ctx.push(
+                    ErrorCode::InvalidArgs,
+                    "expr.args must contain exactly two items",
+                    format!("{}.args", base_path),
+                );
+            }
+        }
+        "slice" => {
+            if !(2..=3).contains(&expr_op.args.len()) {
+                ctx.push(
+                    ErrorCode::InvalidArgs,
+                    "expr.args must contain two or three items",
+                    format!("{}.args", base_path),
+                );
+            }
+        }
+        "zip" => {
+            if expr_op.args.len() < 2 {
+                ctx.push(
+                    ErrorCode::InvalidArgs,
+                    "expr.args must contain at least two items",
+                    format!("{}.args", base_path),
+                );
+            }
+        }
+        "zip_with" => {
+            if expr_op.args.len() < 3 {
+                ctx.push(
+                    ErrorCode::InvalidArgs,
+                    "expr.args must contain at least three items",
+                    format!("{}.args", base_path),
+                );
+            }
+        }
+        "unzip" | "unique" | "sum" | "avg" | "min" | "max" => {
+            if expr_op.args.len() != 1 {
+                ctx.push(
+                    ErrorCode::InvalidArgs,
+                    "expr.args must contain exactly one item",
+                    format!("{}.args", base_path),
+                );
+            }
+        }
+        "sort_by" => {
+            if !(2..=3).contains(&expr_op.args.len()) {
+                ctx.push(
+                    ErrorCode::InvalidArgs,
+                    "expr.args must contain two or three items",
+                    format!("{}.args", base_path),
+                );
+            }
+        }
+        "fold" => {
+            if expr_op.args.len() != 3 {
+                ctx.push(
+                    ErrorCode::InvalidArgs,
+                    "expr.args must contain exactly three items",
+                    format!("{}.args", base_path),
+                );
+            }
+        }
+        "+" | "*" => {
+            if expr_op.args.len() < 2 {
+                ctx.push(
+                    ErrorCode::InvalidArgs,
+                    "expr.args must contain at least two items",
+                    format!("{}.args", base_path),
+                );
+            }
+        }
+        "-" | "/" | "to_base" => {
+            if expr_op.args.len() != 2 {
+                ctx.push(
+                    ErrorCode::InvalidArgs,
+                    "expr.args must contain exactly two items",
+                    format!("{}.args", base_path),
+                );
+            }
+        }
+        "round" => {
+            if !(1..=2).contains(&expr_op.args.len()) {
+                ctx.push(
+                    ErrorCode::InvalidArgs,
+                    "expr.args must contain one or two items",
+                    format!("{}.args", base_path),
+                );
+            }
+        }
+        "date_format" => {
+            if !(2..=4).contains(&expr_op.args.len()) {
+                ctx.push(
+                    ErrorCode::InvalidArgs,
+                    "expr.args must contain two to four items",
+                    format!("{}.args", base_path),
+                );
+            }
+        }
+        "to_unixtime" => {
+            if !(1..=3).contains(&expr_op.args.len()) {
+                ctx.push(
+                    ErrorCode::InvalidArgs,
+                    "expr.args must contain one to three items",
+                    format!("{}.args", base_path),
+                );
+            }
+        }
+        "and" | "or" => {
+            if expr_op.args.len() < 2 {
+                ctx.push(
+                    ErrorCode::InvalidArgs,
+                    "expr.args must contain at least two items",
+                    format!("{}.args", base_path),
+                );
+            }
+        }
+        "not" => {
+            if expr_op.args.len() != 1 {
+                ctx.push(
+                    ErrorCode::InvalidArgs,
+                    "expr.args must contain exactly one item",
+                    format!("{}.args", base_path),
+                );
+            }
+        }
+        "==" | "!=" | "<" | "<=" | ">" | ">=" | "~=" => {
+            if expr_op.args.len() != 2 {
+                ctx.push(
+                    ErrorCode::InvalidArgs,
+                    "expr.args must contain exactly two items",
+                    format!("{}.args", base_path),
+                );
+            }
+        }
+        _ => {}
+    }
+
+    let expr_scope = element_expr_scope(&expr_op.op, false, expr_op.args.len(), scope);
+    for (index, arg) in expr_op.args.iter().enumerate() {
+        let arg_path = format!("{}.args[{}]", base_path, index);
+        let arg_scope = match expr_scope {
+            Some((expr_index, expr_scope)) if expr_index == index => expr_scope,
+            _ => scope,
+        };
+        validate_expr(arg, &arg_path, produced_targets, ctx, arg_scope);
+    }
+}
